backend/dama/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
django-filter>=2.4,<3.0
django-cors-headers>=3.10,<4.0
djangorestframework-simplejwt>=4.8,<5.0
django-import-export>=3.0,<4.0
drf-yasg>=1.20,<2.0
psycopg2>=2.9,<3.0  
python-dotenv>=0.19,<1.0  